move putremoves3 to utils
